test(converter): consolidate helpers at root and group tests by concern
Move every test-only constructor and assertion helper to the tests/mod.rs
root so the submodules carry test functions only, and merge the single-test
files into concern-grouped modules: verdicts, listings, and io.
