+++ solx-benchmark-converter/src/tests/listings.rs
@@ -1,9 +1,25 @@
 //!
-//! Tests for the inline "new failures" regression listing.
+//! Tests for the ranked inline listings: top movers and failure regressions.
 //!
 
+use crate::output::summary::failure_kind::FailureKind;
 use crate::output::summary::failure_regressions::FailureRegression;
 use crate::output::summary::failure_regressions::FailureRegressions;
+use crate::output::summary::top_movers::TopMovers;
+
+#[test]
+fn movers_rank_by_magnitude_regardless_of_direction() {
+    let mut movers = TopMovers::default();
+    movers.push("small", "legacy", 100, 103);
+    movers.push("shrunk", "legacy", 100, 80);
+    movers.push("grown", "legacy", 100, 110);
+    let ranked = movers.ranked();
+    let labels: Vec<&str> = ranked
+        .iter()
+        .map(|movement| movement.label.as_str())
+        .collect();
+    assert_eq!(labels, ["shrunk", "grown", "small"]);
+}
 
 #[test]
 fn failure_regressions_rank_by_magnitude() {
@@ -12,7 +28,7 @@ fn failure_regressions_rank_by_magnitude() {
         regressions.push(FailureRegression {
             label: label.to_owned(),
             mode: "legacy".to_owned(),
-            kind: "test",
+            kind: FailureKind::Test,
             main,
             pr,
         });
